requirements.txt
streamlit
pandas
scikit-learn
fpdf2